Fix displaying speed on GPS data screen.

+++ firmware/sensorio/src/gui/screens/GpsData.c
@@ -48,7 +48,7 @@ static void refresh_task(lv_task_t *p)
   lv_table_set_cell_value_fmt(gpsdata, 3, 1, "%f", data.latitude / 1000000.0);
   lv_table_set_cell_value_fmt(gpsdata, 4, 1, "%f", data.longitude / 1000000.0);
   lv_table_set_cell_value_fmt(gpsdata, 5, 1, "%d", data.numOfSatellites);
-  lv_table_set_cell_value_fmt(gpsdata, 6, 1, "%.03f km/h", data.speed);
+  lv_table_set_cell_value_fmt(gpsdata, 6, 1, "%d km/h", (int)data.speed);
   lv_table_set_cell_value_fmt(gpsdata,
                               7,
                               1,
